[2021-12-2]  final

+++ circle3/philosophers/srcs/utils/ft_get_time.c
@@ -6,7 +6,7 @@
 /*   By: pacman <pacman@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/17 20:18:08 by pacman            #+#    #+#             */
-/*   Updated: 2021/11/27 15:57:57 by pacman           ###   ########.fr       */
+/*   Updated: 2021/12/02 21:25:21 by pacman           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,5 +17,5 @@ long long	ft_get_time(void)
 	static struct timeval	tv;
 
 	gettimeofday(&tv, NULL);
-	return ((tv.tv_sec * (long long)1000) + (tv.tv_usec / 1000));
+	return ((tv.tv_sec * 1000) + (tv.tv_usec / 1000));
 }
